# typeid: improve benchmarks