--- a/typeid/typeid-go/bench_test.go
+++ b/typeid/typeid-go/bench_test.go
@@ -1,3 +1,4 @@
+//nolint:all
 package typeid_test
 
 import (