Update benchmarks to measure more functions and make them comparable with benchstat. Also add benchmarks for vanilla UUIDv7 that can be used as a baseline for parsing/encoding functions.

To measure the impact of a change, run the benchmarks for each commit with:

	go test -count=10 -bench . > (old|new).txt

Then compare them with benchstat:

	go install golang.org/x/perf/cmd/benchstat@latest
	benchstat old.txt new.txt

	# or exclude uuid
	benchstat -filter '-/id:uuid' old.txt new.txt

Initial results:


```
goos: darwin
goarch: arm64
pkg: go.jetpack.io/typeid
                  │     uuid     │                 untyped                  │                  typed                   │
                  │    sec/op    │    sec/op      vs base                   │    sec/op      vs base                   │
New-10               404.6n ± 0%     483.4n ± 1%     +19.47% (p=0.000 n=10)     483.8n ± 1%     +19.55% (p=0.000 n=10)
String/n=1-10        30.29n ± 0%     28.46n ± 0%      -6.01% (p=0.000 n=10)     28.54n ± 0%      -5.78% (p=0.000 n=10)
FromString/n=1-10    24.67n ± 0%     69.88n ± 0%    +183.30% (p=0.000 n=10)     73.11n ± 0%    +196.41% (p=0.000 n=10)
FromBytes/n=1-10     3.118n ± 0%   129.600n ± 0%   +4056.51% (p=0.000 n=10)   130.150n ± 0%   +4074.15% (p=0.000 n=10)
UUIDBytes-10        0.3107n ± 0%   31.9450n ± 0%  +10179.97% (p=0.000 n=10)   32.0900n ± 0%  +10226.63% (p=0.000 n=10)
geomean              12.40n          83.18n         +570.91%                    84.13n         +578.61%

                  │     uuid     │                untyped                │                 typed                 │
                  │     B/op     │    B/op     vs base                   │    B/op     vs base                   │
New-10              16.00 ± 0%     64.00 ± 0%  +300.00% (p=0.000 n=10)     64.00 ± 0%  +300.00% (p=0.000 n=10)
String/n=1-10       48.00 ± 0%     48.00 ± 0%         ~ (p=1.000 n=10) ¹   48.00 ± 0%         ~ (p=1.000 n=10) ¹
FromString/n=1-10    0.00 ± 0%     48.00 ± 0%         ? (p=0.000 n=10)     48.00 ± 0%         ? (p=0.000 n=10)
FromBytes/n=1-10     0.00 ± 0%     96.00 ± 0%         ? (p=0.000 n=10)     96.00 ± 0%         ? (p=0.000 n=10)
UUIDBytes-10         0.00 ± 0%     16.00 ± 0%         ? (p=0.000 n=10)     16.00 ± 0%         ? (p=0.000 n=10)
geomean                        ²   46.88       ?                           46.88       ?
¹ all samples are equal
² summaries must be >0 to compute geomean

                  │     uuid     │                untyped                │                 typed                 │
                  │  allocs/op   │ allocs/op   vs base                   │ allocs/op   vs base                   │
New-10              1.000 ± 0%     3.000 ± 0%  +200.00% (p=0.000 n=10)     3.000 ± 0%  +200.00% (p=0.000 n=10)
String/n=1-10       1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹   1.000 ± 0%         ~ (p=1.000 n=10) ¹
FromString/n=1-10   0.000 ± 0%     2.000 ± 0%         ? (p=0.000 n=10)     2.000 ± 0%         ? (p=0.000 n=10)
FromBytes/n=1-10    0.000 ± 0%     3.000 ± 0%         ? (p=0.000 n=10)     3.000 ± 0%         ? (p=0.000 n=10)
UUIDBytes-10        0.000 ± 0%     1.000 ± 0%         ? (p=0.000 n=10)     1.000 ± 0%         ? (p=0.000 n=10)
geomean                        ²   1.783       ?                           1.783       ?
¹ all samples are equal
² summaries must be >0 to compute geomean
```